Reduce rewrite to filters, move tcph-19 to main test
