+++ tests/query19_fixed.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+#include <vector>
+
 /*
 Query 19 for tpch dataset, manually rewritten:
 select
@@ -55,7 +57,7 @@ where
 */
 
 
-unsigned char __query19_tree_fixed[] = {
+std::vector<unsigned char> __query19_tree_fixed = {
     0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x01, 0x01, 0x1E, 0x1A, 0x00, 0x00,
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x20, 0x02, 0x00, 0x00, 0x00, 0x3C,
     0x3D, 0x09, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00,
