+++ tests/query19_init.h
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <vector>
 /*
 Query 19 for tpch dataset, initial:
 select
@@ -58,7 +59,7 @@ where
 
 
 
-unsigned char __query19_tree_init[] = {
+std::vector<unsigned char> __query19_tree_init = {
     0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x24, 0x01, 0x01, 0x1E, 0x1A, 0x00, 0x00, 0x00,
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x20, 0x01, 0x00, 0x00, 0x00, 0x3D, 0x0C,
     0x00, 0x00, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x02, 0x00,
